pridanie odstranovania tagov a premenovania suborov

+++ src/project/frontend/GalleryPane.java
@@ -1,11 +1,16 @@
 package project.frontend;
 
+import javafx.scene.control.ContextMenu;
+import javafx.scene.control.MenuItem;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.effect.InnerShadow;
+import javafx.scene.input.MouseButton;
+import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.TilePane;
 import javafx.scene.paint.Color;
 import project.backend.Database;
 import project.backend.DatabaseItem;
+import project.backend.SharedBE;
 
 public class GalleryPane extends ScrollPane {
     private static final TilePane tilePane = new TilePane();
@@ -22,23 +27,29 @@ public class GalleryPane extends ScrollPane {
         highlightEffect.setHeight(5);
         highlightEffect.setChoke(1);
 
+        ContextMenu contextMenu = new ContextMenu();
+        MenuItem menuRename = new MenuItem("Rename");
+        menuRename.setOnAction(event -> SharedBE.renameFile(Database.getLastSelectedItem()));
+        contextMenu.getItems().add(menuRename);
+        setContextMenu(contextMenu);
+        setOnContextMenuRequested(event -> contextMenu.show(this, event.getScreenX(), event.getScreenY()));
+
+
         setHbarPolicy(ScrollBarPolicy.NEVER);
         setVbarPolicy(ScrollBarPolicy.AS_NEEDED);
         setFitToWidth(true);
         setContent(tilePane);
     }
 
-    public static void imageViewClicked(DatabaseItem databaseItem) {
+    public static void imageViewClicked(DatabaseItem databaseItem, MouseEvent event) {
         /* assigned in DatabaseItem.setImageView() */
-        Database.setLastSelectedItem(databaseItem);
-        if (!Database.getSelectedItems().contains(databaseItem))
-            Database.addToSelection(databaseItem);
-        else
-            Database.removeIndexFromSelection(databaseItem);
-    }
-
-    public static InnerShadow getHighlightEffect() {
-        return highlightEffect;
+        if (event.getButton().equals(MouseButton.PRIMARY)) {
+            Database.setLastSelectedItem(databaseItem);
+            if (!Database.getSelectedItems().contains(databaseItem))
+                Database.addToSelection(databaseItem);
+            else
+                Database.removeIndexFromSelection(databaseItem);
+        }
     }
 
     public void refreshContent() {
@@ -46,5 +57,9 @@ public class GalleryPane extends ScrollPane {
         for (DatabaseItem item : Database.getFilteredItems()) {
             tilePane.getChildren().add(item.getImageView());
         }
+    }
+
+    public static InnerShadow getHighlightEffect() {
+        return highlightEffect;
     }
 }
