+++ src/project/frontend/LeftPane.java
@@ -1,15 +1,14 @@
 package project.frontend;
 
 import javafx.geometry.Pos;
-import javafx.scene.control.Button;
-import javafx.scene.control.ListCell;
-import javafx.scene.control.ListView;
-import javafx.scene.control.SelectionMode;
+import javafx.scene.control.*;
+import javafx.scene.input.MouseButton;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.paint.Color;
 import project.backend.Database;
 import project.backend.DatabaseItem;
+import project.backend.SharedBE;
 
 import java.util.List;
 
@@ -63,29 +62,39 @@ public class LeftPane extends BorderPane {
                 }
             }
         });
+
+        ContextMenu listContextMenu = new ContextMenu();
+        MenuItem menuRename = new MenuItem("Rename");
+        menuRename.setOnAction(event -> SharedBE.renameFile(Database.getLastSelectedItem()));
+        listContextMenu.getItems().add(menuRename);
+        listView.setContextMenu(listContextMenu);
+        listView.setOnContextMenuRequested(event -> listContextMenu.show(listView, event.getScreenX(), event.getScreenY()));
+
         listView.setOnMouseClicked(event -> {
-            if (displayMode == LeftPaneDisplayMode.NAMES && listView.getSelectionModel().getSelectedItem() != null) {
-                DatabaseItem selectedItem = listView.getSelectionModel().getSelectedItem().getOwner();
-                Database.setLastSelectedItem(selectedItem);
-                if (Database.getSelectedItems().contains(selectedItem))
-                    Database.removeIndexFromSelection(selectedItem);
-                else
-                    Database.addToSelection(selectedItem);
-            } else if (displayMode == LeftPaneDisplayMode.TAGS && listView.getSelectionModel().getSelectedItem() != null) {
-                String tag = listView.getSelectionModel().getSelectedItem().getText();
-                if (whitelist.contains(tag)) {
-                    whitelist.remove(tag);
-                    blacklist.add(tag);
-                    listView.getItems().set(listView.getSelectionModel().getSelectedIndex(), new ColoredText(tag, Color.RED));
-                } else if (blacklist.contains(tag)) {
-                    blacklist.remove(tag);
-                    listView.getItems().set(listView.getSelectionModel().getSelectedIndex(), new ColoredText(tag, Color.BLACK));
-                } else {
-                    whitelist.add(tag);
-                    listView.getItems().set(listView.getSelectionModel().getSelectedIndex(), new ColoredText(tag, Color.GREEN));
+            if (event.getButton().equals(MouseButton.PRIMARY)) {
+                if (displayMode == LeftPaneDisplayMode.NAMES && listView.getSelectionModel().getSelectedItem() != null) {
+                    DatabaseItem selectedItem = listView.getSelectionModel().getSelectedItem().getOwner();
+                    Database.setLastSelectedItem(selectedItem);
+                    if (Database.getSelectedItems().contains(selectedItem))
+                        Database.removeIndexFromSelection(selectedItem);
+                    else
+                        Database.addToSelection(selectedItem);
+                } else if (displayMode == LeftPaneDisplayMode.TAGS && listView.getSelectionModel().getSelectedItem() != null) {
+                    String tag = listView.getSelectionModel().getSelectedItem().getText();
+                    if (whitelist.contains(tag)) {
+                        whitelist.remove(tag);
+                        blacklist.add(tag);
+                        listView.getItems().set(listView.getSelectionModel().getSelectedIndex(), new ColoredText(tag, Color.RED));
+                    } else if (blacklist.contains(tag)) {
+                        blacklist.remove(tag);
+                        listView.getItems().set(listView.getSelectionModel().getSelectedIndex(), new ColoredText(tag, Color.BLACK));
+                    } else {
+                        whitelist.add(tag);
+                        listView.getItems().set(listView.getSelectionModel().getSelectedIndex(), new ColoredText(tag, Color.GREEN));
+                    }
+                    Database.filterByTags();
+                    SharedFE.getGalleryPane().refreshContent();
                 }
-                Database.filterByTags();
-                SharedFrontend.getGalleryPane().refreshContent();
             }
         });
